use proper handler to allow for url verification
