+++ integrations/slack/src/handlers/lens.ts
@@ -1,8 +1,9 @@
 import { Logger } from '@gitbook/runtime';
 
-import { queryLens } from '../actions/queryLens';
+import { queryLens } from '../actions/queryLens'; // eslint-disable-line
 import type { SlashEvent } from '../commands';
 import { SlackRuntimeContext } from '../configuration';
+import { stripBotName } from '../utils';
 
 const logger = Logger('slack:api');
 
@@ -28,4 +29,37 @@ export async function queryLensSlashHandler(slashEvent: SlashEvent, context: Sla
         logger.error('Error calling queryLens. Perhasp no installation was found?');
         return {};
     }
+}
+
+/**
+ * Handle an Event request and route it to the GitBook Lens' query function.
+ */
+export async function queryLensEventHandler(eventPayload: any, context: SlackRuntimeContext) {
+    // pull out required params from the slashEvent for queryLens
+    const { type, text, bot_id, thread_ts, channel, user, team_id } = eventPayload.event;
+
+    // check for bot_id so that the bot doesn't trigger itself
+    if (['message', 'app_mention'].includes(type) && !bot_id) {
+        // strip out the bot-name in the mention and account for user mentions within the query
+        // @ts-ignore
+        const parsedQuery = stripBotName(text, eventPayload.authorizations[0]?.user_id);
+
+        // send to Lens
+        await queryLens({
+            teamId: team_id,
+            channelId: channel,
+            threadId: thread_ts,
+            userId: user,
+            messageType: 'permanent',
+            text: parsedQuery,
+            context,
+            // @ts-ignore
+            authorization: eventPayload.authorizations[0],
+        });
+    }
+
+    // Add custom header(s)
+    return new Response(null, {
+        status: 200,
+    });
 }
